# app/Http/Controllers/GalleryController.php
<?php

namespace App\Http\Controllers;

use Illuminate\Http\Request;
use App\Models\Post;
use Illuminate\Support\Facades\Response; // Add this line   


class GalleryController extends Controller
{
    /**
     * Display a listing of the resource.
     */
    // public function index()
    // {
    //     $data = array(
    //         'id' => "posts",
    //         'menu' => 'Gallery',
    //         'galleries' => Post::where('picture', '!=', 
    //         '')->whereNotNull('picture')->orderBy('created_at', 'desc')->paginate(30));
    //     return view('gallery.index')->with($data);
    // }
    public function index(Request $request)
    {
        $data = [
            'id' => 'posts',
            'menu' => 'Gallery',
            'galleries' => Post::where('picture', '!=', '')
                ->whereNotNull('picture')
                ->orderBy('created_at', 'desc')
                ->paginate(30),
        ];

        if ($request->is('api/*')) {
            // If it's an API request, return JSON response
            return Response::json($data['galleries'], 200);
        } else {
            // If it's a web request, return the view
            return view('gallery.index')->with($data);
        }
    }

    /**
     * Show the form for creating a new resource.
     */
    public function create()
    {
        return view('gallery.create');
    }

    /**
     * Store a newly created resource in storage.
     */
    // public function store(Request $request)
    // {
    //     $this->validate($request, [
    //         'title' => 'required|max:255',
    //         'description' => 'required',
    //         'picture' => 'image|nullable|max:1999'
    //         ]);

    //         if ($request->hasFile('picture')) {
    //             $filenameWithExt = $request->file('picture')->getClientOriginalName();
    //             $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
    //             $extension = $request->file('picture')->getClientOriginalExtension();
    //             $basename = uniqid() . time();
    //             $smallFilename = "small_{$basename}.{$extension}";
    //             $mediumFilename = "medium_{$basename}.{$extension}";
    //             $largeFilename = "large_{$basename}.{$extension}";
    //             $filenameSimpan = "{$basename}.{$extension}";
    //             $path = $request->file('picture')->storeAs('posts_image', $filenameSimpan);
    //         } else {
    //             $filenameSimpan = 'noimage.png';
    //         }

    //         // dd($request->input());
    //         $post = new Post;
    //         $post->picture = $filenameSimpan;
    //         $post->title = $request->input('title');
    //         $post->description = $request->input('description');
    //         $post->save();
    //         return redirect('gallery')->with('success', 'Berhasil menambahkan data baru');
           
    // }

    public function storeApi(Request $request)
    {
        $this->validate($request, [
            'title' => 'required|max:255',
            'description' => 'required',
            'picture' => 'image|nullable|max:1999'
        ]);

        if ($request->hasFile('picture')) {
            $filenameWithExt = $request->file('picture')->getClientOriginalName();
            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
            $extension = $request->file('picture')->getClientOriginalExtension();
            $basename = uniqid() . time();
            $filenameSimpan = "{$basename}.{$extension}";
            $path = $request->file('picture')->storeAs('posts_image', $filenameSimpan);
        } else {
            $filenameSimpan = 'noimage.png';
        }

        $post = new Post;
        $post->picture = $filenameSimpan;
        $post->title = $request->input('title');
        $post->description = $request->input('description');
        $post->save();

        return redirect('gallery')->with('success', 'Berhasil menambahkan data baru');
    }

    /**
     * Display the specified resource.
     */
    public function show(string $id)
    {
        $gallery = Post::findOrFail($id);

        // Return the gallery data in the response
        return Response::json($gallery, 200);
    }

    /**
     * Show the form for editing the specified resource.
     */
    public function edit(string $id)
    {
        $gallery = Post::findOrFail($id);
        return view('gallery.edit', compact('gallery'));
    }

    /**
     * Update the specified resource in storage.
     */
    public function update(Request $request, string $id)
    {
        $this->validate($request, [
            'title' => 'required|max:255',
            'description' => 'required',
            'picture' => 'image|nullable|max:1999'
        ]);
    
        $post = Post::findOrFail($id);
    
        // Handle file update
        if ($request->hasFile('picture')) {
            // Your existing file handling code here
    
            // Delete old image if it's not the default noimage.png
            if ($post->picture !== 'noimage.png') {
                Storage::delete('posts_image/' . $post->picture);
            }
    
            $post->picture = $filenameSimpan;
        }
    
        $post->title = $request->input('title');
        $post->description = $request->input('description');
        $post->save();
    
        return redirect('gallery')->with('success', 'Berhasil memperbarui data');
    }

    /**
     * Remove the specified resource from storage.
     */
    public function destroy(string $id)
    {
        $post = Post::findOrFail($id);
        $post->delete();

        return redirect('gallery')->with('success', 'Berhasil menghapus data');
    }
}