# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contoh API documentation",
        "description": "Contoh API doc menggunakan OpenAPI/Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "choirudin.emchagmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/api/greet": {
            "get": {
                "tags": [
                    "greeting"
                ],
                "summary": "Returns a Sample API response",
                "description": "A sample greeting to test out the API",
                "operationId": "greet",
                "parameters": [
                    {
                        "name": "firstname",
                        "in": "query",
                        "description": "nama depan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "description": "nama belakang",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/gallery": {
            "get": {
                "tags": [
                    "gallery"
                ],
                "summary": "Get gallery data",
                "description": "Retrieve gallery data",
                "operationId": "gallery.index",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "gallery"
                ],
                "summary": "Create a new post with an image",
                "description": "Create a new post with an image in the gallery",
                "operationId": "gallery.storeApi",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "picture"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Post title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Post description",
                                        "type": "string"
                                    },
                                    "picture": {
                                        "description": "Post image",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        }
    }
}